[#544] Modularize logic classes for clarity
- Extracted logic classes into separate files for better organization
- Simplified `logic.rb` by moving `Authentication`, `Dashboard`,
  `Incoming`, `Misc`, `Secrets` modules to their respective files
- Improved readability and maintainability by grouping related logic
- Kept necessary `requires` to ensure all dependencies are met
