+++ lib/onetime/logic/welcome.rb
@@ -151,108 +151,3 @@ module Onetime::Logic
 
   end
 end
-
-__END__
-{
-  "id": "cs_test_a1rXKe8udp",
-  "object": "checkout.session",
-  "after_expiration": null,
-  "allow_promotion_codes": false,
-  "amount_subtotal": 2500,
-  "amount_total": 2500,
-  "automatic_tax": {
-    "enabled": true,
-    "liability": {
-      "type": "self"
-    },
-    "status": "complete"
-  },
-  "billing_address_collection": "auto",
-  "cancel_url": "https://stripe.com",
-  "client_reference_id": "",
-  "client_secret": null,
-  "consent": null,
-  "consent_collection": {
-    "payment_method_reuse_agreement": null,
-    "promotions": "none",
-    "terms_of_service": "none"
-  },
-  "created": 1722978823,
-  "currency": "usd",
-  "currency_conversion": null,
-  "custom_fields": [],
-  "custom_text": {
-    "after_submit": null,
-    "shipping_address": null,
-    "submit": null,
-    "terms_of_service_acceptance": null
-  },
-  "customer": "cus_Qc8zvS3zCmHusS",
-  "customer_creation": "if_required",
-  "customer_details": {
-    "address": {
-      "city": null,
-      "country": "CA",
-      "line1": null,
-      "line2": null,
-      "postal_code": "H2W 1Y7",
-      "state": null
-    },
-    "email": "gregproops@example.com",
-    "name": "Greg Proops",
-    "phone": null,
-    "tax_exempt": "none",
-    "tax_ids": []
-  },
-  "customer_email": null,
-  "expires_at": 1723065223,
-  "invoice": "in_1PkuhzHA8OZxV3C",
-  "invoice_creation": null,
-  "livemode": false,
-  "locale": "auto",
-  "metadata": {},
-  "mode": "subscription",
-  "payment_intent": null,
-  "payment_link": "plink_1PMwedHA8OZxV3Cy",
-  "payment_method_collection": "always",
-  "payment_method_configuration_details": null,
-  "payment_method_options": {
-    "card": {
-      "request_three_d_secure": "automatic"
-    }
-  },
-  "payment_method_types": [
-    "card"
-  ],
-  "payment_status": "paid",
-  "phone_number_collection": {
-    "enabled": false
-  },
-  "recovered_from": null,
-  "saved_payment_method_options": {
-    "allow_redisplay_filters": [
-      "always"
-    ],
-    "payment_method_remove": null,
-    "payment_method_save": null
-  },
-  "setup_intent": null,
-  "shipping_address_collection": null,
-  "shipping_cost": null,
-  "shipping_details": null,
-  "shipping_options": [],
-  "status": "complete",
-  "submit_type": "auto",
-  "subscription": "sub_1PkuhzHA8OZxV3CzA",
-  "success_url": "https://staging.onetimesecret.com/welcome?checkout={CHECKOUT_SESSION_ID}",
-  "tax_id_collection": {
-    "enabled": true
-  },
-  "total_details": {
-    "amount_discount": 0,
-    "amount_shipping": 0,
-    "amount_tax": 0
-  },
-  "ui_mode": "hosted",
-  "url": null
-}
